增加了Token 存入 Redis 和一个新的CheckToken工具类

+++ src/main/java/edu/lgxy/lbj/asms/controller/AdminController.java
@@ -1,5 +1,6 @@
 package edu.lgxy.lbj.asms.controller;
 
+import edu.lgxy.lbj.asms.config.CheckToken;
 import edu.lgxy.lbj.asms.config.JsonResult;
 import edu.lgxy.lbj.asms.config.Page;
 import edu.lgxy.lbj.asms.config.PageAndUserId;
@@ -9,11 +10,14 @@ import edu.lgxy.lbj.asms.qo.PageQo2;
 import edu.lgxy.lbj.asms.qo.UserPage;
 import edu.lgxy.lbj.asms.service.*;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.sql.ResultSet;
 import java.util.HashMap;
 import java.util.List;
@@ -22,6 +26,12 @@ import java.util.Map;
 @Slf4j
 @RestController
 public class AdminController {
+    @Autowired
+    private HttpServletRequest request;
+    @Autowired
+    private HttpSession session;
+    @Autowired
+    private CheckToken checkToken;
     @Resource
     private UserService userService;
     @Resource
@@ -35,6 +45,12 @@ public class AdminController {
     //删除竞赛信息
     @RequestMapping("/deleteContestInfo")
     JsonResult<Map> deleteContestInfo(@RequestBody ContestInformation contestInformation){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         log.info("%%%%%%%%%%%%%%%"+contestInformation);
         int result = contestService.deleteContestInfo(contestInformation);
         if(result<=0){
@@ -45,6 +61,12 @@ public class AdminController {
     //新增竞赛信息
     @RequestMapping("/insertContestInfo")
     JsonResult<Map> insertContestInfo(@RequestBody ContestInformation contestInformation){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         log.info("%%%%%%%%%%%%%%%"+contestInformation);
         int result = contestService.insertContestInfo(contestInformation);
         if(result<=0){
@@ -55,7 +77,12 @@ public class AdminController {
     //获取 所有申报的项目
     @RequestMapping("/getDeclarationListByPage")
     JsonResult<Map> getDeclarationListByPage(PageQo2 pageQo){
-
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         int pageSize=pageQo.getPageSize();
         int pageIndex= pageQo.getPageIndex();
         int applicantId = pageQo.getApplicantId();
@@ -85,6 +112,12 @@ public class AdminController {
     //获取报名人信息
     @RequestMapping("/admin/getParticipantListByPage")
     JsonResult<Map> getParticipantListByPage(PageQo pageQo){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         int pageSize=pageQo.getPageSize();
         int pageIndex= pageQo.getPageIndex();
         String contestDate=pageQo.getContestDate();
@@ -112,6 +145,12 @@ public class AdminController {
     //获取报名人成绩信息
     @RequestMapping("/admin/gradesManagement")
     JsonResult<Map> gradesManagement(PageQo pageQo){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         int pageSize=pageQo.getPageSize();
         int pageIndex= pageQo.getPageIndex();
         String contestDate=pageQo.getContestDate();
@@ -140,6 +179,12 @@ public class AdminController {
     //更新已发布的竞赛信息
     @RequestMapping("/updateDeclaration")
     JsonResult<Map> updateDeclaration(@RequestBody ContestDeclaration contestDeclaration){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         Map<String,Object> map = new HashMap<>();
         String msg="";
         String code="";
@@ -162,6 +207,12 @@ public class AdminController {
 
     @RequestMapping("/getUserInformation")
     JsonResult<Map> getUserInformation(UserPage userPage){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         int pageSize=userPage.getPageSize();
         int pageIndex= userPage.getPageIndex();
         String userGroup= userPage.getUserGroup();
@@ -186,6 +237,12 @@ public class AdminController {
 
     @RequestMapping("/updateUserInformation")
     JsonResult<Map> updateUserInformation(@RequestBody User user){
+        String username = request.getHeader("username");
+        String token = request.getHeader("token");
+        JsonResult<Map> jsonResult= checkToken.checkTokenByUserName(username,token);
+        if(jsonResult!=null){
+            return jsonResult;
+        }
         log.info("!!!!!!!!!!!!!!!"+user);
         int result = userService.updateUserInformation(user);
         if(result<=0){
